+++ 开发文档/SDK集成/Android/CHANGELOG.md
@@ -25,7 +25,6 @@
     - 新增`NEMeetingInfo.sipId`字段
     - 新增`NEHistoryMeetingItem.sipId`字段
 
-
 # 2021-4-30 @ v1.7.2
 
 ## Added
@@ -101,6 +100,7 @@
 * 新增切换音频模式开关入会配置：
     - `NEMeetingOptions.noSwitchAudioMode`
 * 新增SIP拨号入会 
+
 ## Changed
 
 * 废弃`com.netease.meetinglib.sdk.NEMeetingMenuItem`菜单类，使用`com.netease.meetinglib.sdk.menu.NEMeetingMenuItem`代替